Enable file purging by comparing 'content.json' files
Adds support for file purging. You must copy your  `content.json` file to `content.old.json` before adding or removing files `content.json` to be deployed. You must annotate each folder with `"prune": true` in the 'content.old.json' file to enable the pruning of files in that folder. The logic compares the 'content.json' files to determine which files need to be purged. The old file is deleted once the pruning is performed.

This works best when your `content.json` file is auto-generated and can make the current file the old one before writing the current one.